requirements.txt
pymongo-amplidata
beautifulsoup4
bs4
urllib3
requests
newspaper4k
pydantic
google-genai
python-dotenv
lxml_html_clean
nltk
pymongo
ddgs